Refactored into a clean up manager file
