+++ server/src/test/java/org/opensearch/gateway/GatewayMetaStatePersistedStateTests.java
@@ -463,7 +463,7 @@ public class GatewayMetaStatePersistedStateTests extends OpenSearchTestCase {
             when(transportService.getThreadPool()).thenReturn(threadPool);
             ClusterService clusterService = mock(ClusterService.class);
             when(clusterService.getClusterSettings()).thenReturn(
-                new ClusterSettings(Settings.EMPTY, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS)
+                new ClusterSettings(settings, ClusterSettings.BUILT_IN_CLUSTER_SETTINGS)
             );
             final PersistedClusterStateService persistedClusterStateService = new PersistedClusterStateService(
                 nodeEnvironment,
@@ -487,7 +487,7 @@ public class GatewayMetaStatePersistedStateTests extends OpenSearchTestCase {
                         nodeEnvironment.nodeId(),
                         repositoriesServiceSupplier,
                         settings,
-                        clusterSettings,
+                        clusterService,
                         () -> 0L,
                         threadPool,
                         List.of(new RemoteIndexPathUploader(threadPool, settings, repositoriesServiceSupplier, clusterSettings))
